Added new configurations or hdx custom exports and initial code for hdx table
