+++ .github/workflows/Unit-Test.yml
@@ -33,6 +33,7 @@ jobs:
       ALLOW_BIND_ZIP_FILTER: ${{ secrets.ALLOW_BIND_ZIP_FILTER }}
       APP_SECRET_KEY: ${{ secrets.APP_SECRET_KEY }}
       ENABLE_HDX_EXPORTS: ${{ secrets.ENABLE_HDX_EXPORTS }}
+      ENABLE_CUSTOM_EXPORTS: True
       ENABLE_POLYGON_STATISTICS_ENDPOINTS: ${{ secrets.ENABLE_POLYGON_STATISTICS_ENDPOINTS }}
       POLYGON_STATISTICS_API_URL: ${{ secrets.POLYGON_STATISTICS_API_URL }}
       HDX_API_KEY: ${{ secrets.HDX_API_KEY }}
