+++ src/config.py
@@ -195,6 +195,16 @@ USE_DUCK_DB_FOR_CUSTOM_EXPORTS = os.environ.get(
     "USE_DUCK_DB_FOR_CUSTOM_EXPORTS"
 ) or config.getboolean("API_CONFIG", "USE_DUCK_DB_FOR_CUSTOM_EXPORTS", fallback=False)
 
+if USE_DUCK_DB_FOR_CUSTOM_EXPORTS:
+    DUCK_DB_MEMORY_LIMIT = os.environ.get("DUCK_DB_MEMORY_LIMIT") or config.get(
+        "API_CONFIG", "DUCK_DB_MEMORY_LIMIT", fallback=None
+    )
+    DUCK_DB_THREAD_LIMIT = os.environ.get("DUCK_DB_THREAD_LIMIT") or config.get(
+        "API_CONFIG", "DUCK_DB_THREAD_LIMIT", fallback=None
+    )
+ENABLE_CUSTOM_EXPORTS = os.environ.get("ENABLE_CUSTOM_EXPORTS") or config.getboolean(
+    "API_CONFIG", "ENABLE_CUSTOM_EXPORTS", fallback=False
+)
 
 HDX_SOFT_TASK_LIMIT = os.environ.get("HDX_SOFT_TASK_LIMIT") or config.get(
     "HDX", "HDX_SOFT_TASK_LIMIT", fallback=5 * 60 * 60
@@ -270,13 +280,6 @@ if ENABLE_HDX_EXPORTS:
             else None
         )
     )
-    if USE_DUCK_DB_FOR_CUSTOM_EXPORTS:
-        DUCK_DB_MEMORY_LIMIT = os.environ.get("DUCK_DB_MEMORY_LIMIT") or config.get(
-            "HDX", "DUCK_DB_MEMORY_LIMIT", fallback=None
-        )
-        DUCK_DB_THREAD_LIMIT = os.environ.get("DUCK_DB_THREAD_LIMIT") or config.get(
-            "HDX", "DUCK_DB_THREAD_LIMIT", fallback=None
-        )
 
 
 def get_db_connection_params() -> dict:
